==> 1.EncodingProcessing.ipynb <==
import glob
import pandas as pd
import numpy as np
import xml.etree.ElementTree as et
import os
from datetime import datetime
import re
from tqdm import tqdm

#Opening out stored filenames and their encodings
with open('Blogs/output.txt','r') as f:
    file =f.read()

data = list(re.split('\n|: |; | |charset=',file))

#removing empty elements in list
while '' in data:
    data.remove('')

len(data)

df = pd.DataFrame(columns='Filename Type Encoding'.split())

df

i=0
while i<len(data):
    filename = data[i]
    i+=1
    type_data = data[i]
    i+=1
    encoding = data[i]
    i+=1
    dict_data = {'Filename':filename,'Type':type_data,'Encoding':encoding}
    df = df.append(dict_data,ignore_index=True)

df.drop(19320,axis=0,inplace=True)

df

df['Type'].value_counts()

df['Encoding'].value_counts()

df[df['Type']=='application/octet-stream']

## The above data shows us that all binary encoding files are basically of the type: application/octet-stream


df[df['Type']=='text/x-c']

## We can see from the above that type text/x-c supports utf-8

### ----------------------------------------------------------------------------------------------------------------------------


## Side Note: Since I had trouble opening the files with different encodings using xml.ElementTree I decided to process files separately according to their encoding type to make sure all the data gets correctly imported

### ----------------------------------------------------------------------------------------------------------------------------


# Process: Importing data from filetype and storing them as a dataframe csv file

## 1. ASCII Filenames

filenames_ascii = list(df[df['Encoding']=='us-ascii']['Filename'])

len(filenames_ascii)

os.chdir('/Users/deepakbuddha/Desktop/scoutbee_proj/blogs')

xmldata_ascii = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

xmldata_ascii

# Pattern format for the date column
pattern = r"\d{2},\w+,\d{4}"

for file in tqdm(filenames_ascii,desc='Progress'):
    blog = ''
    
    #using errors = "replace" rather than "surrogateescape" as with "replace" we can export data into csv file
    with open(file,errors="replace") as f:
        for line in f:
            blog = blog + line
    
    #removing all the tags
    tempdata = list(re.split('<Blog>|<date>|</date>|<post>|</post>|</Blog>',blog))
    
    #loop to remove unwanted elements which just contain space characters (1 or more) from our list
    blogdata = []
    for ele in tempdata:
        if ele.strip():
            blogdata.append(ele)
            
     #getting the Id, Gender, Age, Profession and AstroSign of that particular blogger
    userid = file.split('.')[0]
    gender = file.split('.')[1]
    age = file.split('.')[2]
    field_of_work = file.split('.')[3]
    ast_sign = file.split('.')[4]

    i=0
    for i in range(len(blogdata)):
        # If our list element contains date format and has less than 20 characters
        # We use less than 20 characters to make sure to omit considering date format from our blogposts
        if ((len(blogdata[i])<20) and (len(re.findall(pattern,blogdata[i]))!=0)):
            date = blogdata[i]
        else:
            post = blogdata[i]
            
            #Since a typical blog post consists of a pair of date and post, we append the values after getting our post
            blog_dict ={'Id':userid,'Date':date, 'Post':post, 'Gender':gender,'Age':age,'Industry':field_of_work,'Astrosign':ast_sign}
            xmldata_ascii = xmldata_ascii.append(blog_dict,ignore_index=True)

xmldata_ascii

xmldata_ascii.to_csv('../CSVData/ASCIIData.csv',index=False)

### ----------------------------------------------------------------------------------------------------------------------------


## 2. UTF8 Filenames

filenames_utf8 = list(df[df['Encoding']=='utf-8']['Filename'])

len(filenames_utf8)

xmldata_utf8 = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

xmldata_utf8

# Pattern format for the date column
pattern = r"\d{2},\w+,\d{4}"

for file in tqdm(filenames_utf8,desc='Progress'):
    blog = ''
    
    #using errors = "replace" rather than "surrogateescape" as with "replace" we can export data into csv file
    with open(file,errors="replace") as f:
        for line in f:
            blog = blog + line
    
    #removing all the tags
    tempdata = list(re.split('<Blog>|<date>|</date>|<post>|</post>|</Blog>',blog))
    
    #loop to remove unwanted elements which just contain space characters (1 or more) from our list
    blogdata = []
    for ele in tempdata:
        if ele.strip():
            blogdata.append(ele)
            
    #getting the Id, Gender, Age, Profession and AstroSign of that particular blogger
    userid = file.split('.')[0]
    gender = file.split('.')[1]
    age = file.split('.')[2]
    field_of_work = file.split('.')[3]
    ast_sign = file.split('.')[4]

    i=0
    for i in range(len(blogdata)):
        # If our list element contains date format and has less than 20 characters
        # We use less than 20 characters to make sure to omit considering date format from our blogposts
        if ((len(blogdata[i])<20) and (len(re.findall(pattern,blogdata[i]))!=0)):
            date = blogdata[i]
        else:
            post = blogdata[i]
            
            #Since a typical blog post consists of a pair of date and post, we append the values after getting our post
            blog_dict ={'Id':userid,'Date':date, 'Post':post, 'Gender':gender,'Age':age,'Industry':field_of_work,'Astrosign':ast_sign}
            xmldata_utf8 = xmldata_utf8.append(blog_dict,ignore_index=True)

xmldata_utf8

xmldata_utf8['Id'].nunique()

xmldata_utf8.to_csv('../CSVData/UTF8Data.csv',index=False)

### ----------------------------------------------------------------------------------------------------------------------------


##  3. iso-8859-1 Filenames

filenames_iso = list(df[df['Encoding']=='iso-8859-1']['Filename'])

len(filenames_iso)

xmldata_iso = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

xmldata_iso

pattern = r"\d{2},\w+,\d{4}"

for file in tqdm(filenames_iso,desc='Progress'):
    blog = ''
    with open(file,errors="replace") as f:
        for line in f:
            blog = blog + line
    
    tempdata = list(re.split('<Blog>|<date>|</date>|<post>|</post>|</Blog>',blog))
    
    blogdata = []
    for ele in tempdata:
        if ele.strip():
            blogdata.append(ele)
            
    userid = file.split('.')[0]
    gender = file.split('.')[1]
    age = file.split('.')[2]
    field_of_work = file.split('.')[3]
    ast_sign = file.split('.')[4]

    i=0
    for i in range(len(blogdata)):
        if ((len(blogdata[i])<20) and (len(re.findall(pattern,blogdata[i]))!=0)):
            date = blogdata[i]
        else:
            post = blogdata[i]
            blog_dict ={'Id':userid,'Date':date, 'Post':post, 'Gender':gender,'Age':age,'Industry':field_of_work,'Astrosign':ast_sign}
            xmldata_iso = xmldata_iso.append(blog_dict,ignore_index=True)

xmldata_iso

xmldata_iso['Id'].nunique()

xmldata_iso.to_csv('../CSVData/ISOData.csv',index=False)

### ----------------------------------------------------------------------------------------------------------------------------


## 4.  binary  Filenames

filenames_bin = list(df[df['Encoding']=='binary']['Filename'])

len(filenames_bin)

xmldata_bin = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

xmldata_bin

pattern = r"\d{2},\w+,\d{4}"

for file in tqdm(filenames_bin,desc='Progress'):
    blog = ''
    with open(file,errors="replace") as f:
        for line in f:
            blog = blog + line
    
    tempdata = list(re.split('<Blog>|<date>|</date>|<post>|</post>|</Blog>',blog))
    
    blogdata = []
    for ele in tempdata:
        if ele.strip():
            blogdata.append(ele)
            
    userid = file.split('.')[0]
    gender = file.split('.')[1]
    age = file.split('.')[2]
    field_of_work = file.split('.')[3]
    ast_sign = file.split('.')[4]

    i=0
    for i in range(len(blogdata)):
        if ((len(blogdata[i])<20) and (len(re.findall(pattern,blogdata[i]))!=0)):
            date = blogdata[i]
        else:
            post = blogdata[i]
            blog_dict ={'Id':userid,'Date':date, 'Post':post, 'Gender':gender,'Age':age,'Industry':field_of_work,'Astrosign':ast_sign}
            xmldata_bin = xmldata_bin.append(blog_dict,ignore_index=True)

xmldata_bin

xmldata_bin['Id'].nunique()

xmldata_bin.to_csv('../CSVData/BINData.csv',index=False)

### ----------------------------------------------------------------------------------------------------------------------------


## 5.  unknown-8bit Filenames

filenames_unknown = list(df[df['Encoding']=='unknown-8bit']['Filename'])

len(filenames_unknown)

xmldata_unknown = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

xmldata_unknown

pattern = r"\d{2},\w+,\d{4}"

for file in tqdm(filenames_unknown,desc='Progress'):
    blog = ''
    with open(file,errors="replace") as f:
        for line in f:
            blog = blog + line
    
    tempdata = list(re.split('<Blog>|<date>|</date>|<post>|</post>|</Blog>',blog))
    
    blogdata = []
    for ele in tempdata:
        if ele.strip():
            blogdata.append(ele)
            
    userid = file.split('.')[0]
    gender = file.split('.')[1]
    age = file.split('.')[2]
    field_of_work = file.split('.')[3]
    ast_sign = file.split('.')[4]

    i=0
    for i in range(len(blogdata)):
        if ((len(blogdata[i])<20) and (len(re.findall(pattern,blogdata[i]))!=0)):
            date = blogdata[i]
        else:
            post = blogdata[i]
            blog_dict ={'Id':userid,'Date':date, 'Post':post, 'Gender':gender,'Age':age,'Industry':field_of_work,'Astrosign':ast_sign}
            xmldata_unknown = xmldata_unknown.append(blog_dict,ignore_index=True)

xmldata_unknown

xmldata_unknown['Id'].nunique()

xmldata_unknown.to_csv('../CSVData/Unknown8bitData.csv',index=False)
