==> 2.DataMerge.ipynb <==
import glob
import pandas as pd
import numpy as np
import xml.etree.ElementTree as et
import os
from datetime import datetime
import re
from tqdm import tqdm

# Creating a dataframe with our required fields
df = pd.DataFrame(columns='Id Date Post Gender Age Industry Astrosign'.split())

# Importing ASCII Data
df1 = pd.read_csv('CSVData/ASCIIData.csv')

# Importing UTF-8 Data
df2 = pd.read_csv('CSVData/UTF8Data.csv')

# Importing BIN Data
df3 = pd.read_csv('CSVData/BINData.csv')

# Importing unknown-8bit Data
df4 = pd.read_csv('CSVData/Unknown8bitData.csv')

# Importing ISO-8859-1 Data
df5 = pd.read_csv('CSVData/ISOData.csv')

df1.shape

df2.shape

df3.shape

df4.shape

df5.shape

datas = [df1,df2,df3,df4,df5]

# Our combined data across all the blog files
df = pd.concat(datas)

df.shape

df.to_csv('CSVData/Allformatdata.csv',index=False)

df
