==> 3.PipelineNLP.ipynb <==
# DISCLAIMER: Please read the file before you run it because the tokenizer takes about 6 hours to run

# Also FYI, the Search Query Input is at the end of the Notebook (before Conclusion)

### Note:

#### You can find the tokenized data at: CSVData/Tokenized_Allformatdata.csv
#### Vocabulary at: CSVData/Vocab_token.csv
#### TF-IDF of the blog at: Pickle/blog_vector_tfidf.pickle

# Now lets dive into the code itself

import glob
import pandas as pd
import numpy as np
import xml.etree.ElementTree as et
import os
from datetime import datetime
import re
from tqdm import tqdm
from scipy import spatial
import spacy
import string
from sklearn.metrics.pairwise import cosine_similarity


#Reading our processed data from all csv files
df = pd.read_csv('CSVData/Allformatdata.csv')

df.shape

df.head()

#Checking if all the files have been included in our df
df['Id'].nunique()

#Trying to fetch a file and randomly checking out parameters
print(df[df['Id']==813360]['Post'])

print(df[df['Id']==813360]['Post'][645276])

## ---------------------------------------------------------------------------------------------------------

## Tokenizer (tokenize, lemmatize and remove stop words

from spacy.lang.en.stop_words import STOP_WORDS

spacy_nlp = spacy.load('en_core_web_lg')

#create list of punctuations and stopwords
punctuations = string.punctuation
stop_words = spacy.lang.en.stop_words.STOP_WORDS

#function for data cleaning and processing
#This can be further enhanced by adding / removing reg-exps as desired.

def spacy_tokenizer(sentence):
 
    #remove distracting single quotes
    sentence = re.sub('\'','',sentence)

    #remove digits and words containing digits
    sentence = re.sub('\w*\d\w*','',sentence)

    #replace extra spaces with single space
    sentence = re.sub(' +',' ',sentence)

    #remove unwanted lines starting from special charcters
    sentence = re.sub(r'\n: \'\'.*','',sentence)
    sentence = re.sub(r'\n!.*','',sentence)
    sentence = re.sub(r'^:\'\'.*','',sentence)
    
    #remove non-breaking new line characters
    sentence = re.sub(r'\n',' ',sentence)
    
    #remove punctunations
    sentence = re.sub(r'[^\w\s]',' ',sentence)
    
    #creating token object
    tokens = spacy_nlp(sentence)
    
    #lower, strip and lemmatize
    tokens = [word.lemma_.lower().strip() if word.lemma_ != "-PRON-" else word.lower_ for word in tokens]
    
    #remove stopwords, and exclude words less than 2 characters
    tokens = [word for word in tokens if word not in stop_words and word not in punctuations and len(word) > 2]
    
    #return tokens
    return tokens

df

#Using tqdm to see the progress of running a file
tqdm.pandas()

#Tokenize our ENTIRE Dataset
df['Post_tokenized'] = df['Post'].progress_map(lambda x: spacy_tokenizer(x))

### Uff, that took 6 hours!!

## ---------------------------------------------------------------------------------------------------------

## Section 1: TF-IDF Vectorizer

def identity_tokenizer(text):
    return text

from sklearn.feature_extraction.text import TfidfVectorizer
import operator

## Create Vocabulary
vocabulary = set()

for doc in df.Post_tokenized:
        vocabulary.update(doc)
        
vocabulary = list(vocabulary)

# Intializating the tfIdf model
tfidf = TfidfVectorizer(tokenizer=identity_tokenizer,vocabulary=vocabulary,lowercase=False)

# Fit the TfIdf model
blog_vector = tfidf.fit_transform(df.Post_tokenized)


blog_vector

len(vocabulary)

blog_vector.shape

## ---------------------------------------------------------------------------------------------------------

## Section 2: Getting similar words for our query from Spacy Large Vocab library

nlp = spacy.load('en_core_web_lg')

def createKeywordsVectors(keyword, nlp):
    doc = nlp(keyword)  # convert to document object

    return doc.vector



# method to find cosine similarity using Spacy and Spatial distance
def cosineSimilarity(vect1, vect2):
    # return cosine distance
    return 1 - spatial.distance.cosine(vect1, vect2)


# method to find similar words
def getSimilarWords(keyword, nlp):
    similarity_list = []

    keyword_vector = createKeywordsVectors(keyword, nlp)

    for tokens in nlp.vocab:
        if (tokens.has_vector):
            if (tokens.is_lower):
                if (tokens.is_alpha):
                    similarity_list.append((tokens, cosineSimilarity(keyword_vector, tokens.vector)))

    similarity_list = sorted(similarity_list, key=lambda item: -item[1])
    similarity_list = similarity_list[:30]

    top_similar_words = [item[0].text for item in similarity_list]

    top_similar_words = top_similar_words[:10]
    top_similar_words.append(keyword)

    for token in nlp(keyword):
        top_similar_words.insert(0, token.lemma_)

    for words in top_similar_words:
        if words.endswith("s"):
            top_similar_words.append(words[0:len(words)-1])

    top_similar_words = list(set(top_similar_words))

    top_similar_words = [words for words in top_similar_words]
    
    stop_words = spacy.lang.en.stop_words.STOP_WORDS
    
    #remove stopwords, and exclude words less than 2 characters
    top_similar_words = [word for word in top_similar_words if word not in stop_words and word not in punctuations and len(word) > 2]


    return ", ".join(top_similar_words)

#Taking in our tokenized query entry and finding similar matches
def Similarkey(query_token):
    
    #Finding all similar words
    similarkey = []
    for i in query_token:
        #Getting similar words
        result = getSimilarWords(i,nlp)
        result = result.split(', ')
        for el in result:
            similarkey.append(el)
            
    #Joining our results with our query_token        
    joinedkey = similarkey + query_token
    
    #Removing duplicates in our joinedkey
    j_key = set(joinedkey)
    j_key = list(joinedkey)
    
    #Returning similar words along withour query tokens as a list
    return(j_key)

## ---------------------------------------------------------------------------------------------------------

## Our SemanticSearch combining similarity of words and feeding it into a TF-IDF vectorizer

## Basically Section 2 -- > Section 1 = Our Semantic Search

def SemanticSearch(word_input,x):
    #Creating a dataframe to store the token of the word_input as list
    query_df = pd.DataFrame(columns=['token'])    

    query_token = spacy_tokenizer(word_input)
    similar_token = Similarkey(query_token)
    query_df.loc[0,'token'] = query_token

    #Transforming our input query as a TF IDF vector
    input_vector = tfidf.transform(query_df['token'])   

    #Getting cosine similarities
    cosines = cosine_similarity(input_vector, blog_vector).flatten()

    #Ordering our results in descending order
    out = np.array(cosines).argsort()[-x:][::-1]

    #Seeing the results of our query in our original dataframe
    result_df = df.loc[out]
    result_df['Relevance'] = cosines[out]
    return(result_df)
    
    

## ---------------------------------------------------------------------------------------------------------

## Saving the Tokenized Data, Vocabulary and also storing our blog_vector (from TDFIDF) along with Vocabulary (again) 

df.to_csv('CSVData/Tokenized_Allformatdata.csv',index=False)

vocab_df = pd.DataFrame(vocabulary)

vocab_df.to_csv('CSVData/Vocab_token.csv',index=False)


import pickle

pickle.dump(blog_vector, open("Pickle/blog_vector_tfidf.pickle", "wb"))

pickle.dump(vocabulary,open("Pickle/vocabulary.pickle","wb"))

## ---------------------------------------------------------------------------------------------------------

# FINALLY HERE IS OUR SEARCH ENGINE!

print("Welcome to the Semantic Search Engine!\n\nKindly enter the keyword/sentence/query you are looking for:\t")
word = input()
print("\n How many relevant results who you like me to show?:\t")
r = input()
r = int(r)

result_df = SemanticSearch(word,r)
display(result_df)


## Another run

print("Welcome to the Semantic Search Engine!\n\nKindly enter the keyword/sentence/query you are looking for:\t")
word = input()
print("\n How many relevant results who you like me to show?:\t")
r = input()
r = int(r)

result_df = SemanticSearch(word,r)
display(result_df)

## One more final run

print("Welcome to the Semantic Search Engine!\n\nKindly enter the keyword/sentence/query you are looking for:\t")
word = input()
print("\n How many relevant results who you like me to show?:\t")
r = input()
r = int(r)

result_df = SemanticSearch(word,r)
display(result_df)

# CONCLUSIONS


## 1. Making use of POS and NER

### The above cell output from our semantic search is of particular interest to us. We can see that when we expand the relevant results to a higher number (20 in this case) we see that in some cases the word nice seems to get a higher weightage while our sentence subject was "weather" . 

### We could use Part of Speech in this case and give more weightage to "weather" than "nice". In other cases, we can make sure of Name Entity Recognition and give more weight to those words.

### Given a week time to accomplish the task, and to maximize the output:efficiency ratio, I had decided to omit those two parts from my pipeline.

..

## 2. Utlitizing Date column to sort our search based on recent/older posts

## Another factor on which we can model is sorting by converting the date column to a datetime object and sorting based on recent/ older results. The hurdle which I encountered while trying to do this was that the date column has entries in the month column in different languages( Eg., Romanian, Italian, French.. to name a few)

## For more details checkout Scrapdrafts/DetectLanguagesinDates.ipynb for validation of the above statement.

..

## 3. Topic Modelling

### Finally we could potentially topic model our data based on Gender, Industry, Age (AstroSign could also possibly be included, though that would be a stretch :D) to make note of the jargon that particular group of people use and give it more weight.

# Hope you enjoyed going through the code. Any inputs on improving the comments, readability or the code itself are most welcome :)
